code/requirements.txt
RPI.GPIO
rpi_ws281x
adafruit-circuitpython-neopixel
azure-iot-device
python-dotenv